Message of the day plugin: initial commit
This plugin provides a convenient way for Gerrit admins to publish a
message of the day to Gerrit users.

Besides implementing the MessageOfTheDay extension point this plugin
also provides a concept of message expiration date which ensures
that the current message is shown only until, and including, the
expiration date.

Change-Id: I6adc747389d797de1888221bc0bf8ffe64865c26