Implement the MessageOfTheDay functionality completely in the plugin
The MessageOfTheDay extension point was removed from Gerrit core. Thus,
this plugin was not working anymore with Gerrit version 3.0 or higher.

This change moves the complete handling of the messages to the plugin.
The messages are still configured in the
$GERRIT_SITE/etc/messageoftheday.config-file and are given in html-files
in the $GERRIT_SITE/data/messageoftheday directory. The latest configured
message will be provided by a REST API endpoint implemented by the plugin.
A UI element using the 'banner' endpoint will show the message as provided
by the REST API.

Change-Id: I05f7e7c672389f809ee6ea143656d5335dd8ffe0