Fix redisplay time zone issue
Before this change, the client was not converting the redisplay time
to its local time. Due to this issue, when the client is in IST and
server sends UTC time in redisplay, then the banner is re-displayed to
the user much sooner than intended.

Moreover, the redisplay format "Aug 29, 2023, 12:00:00 AM" is not
supported by the Set-Cookie expires field, where it was ignoring
AM and PM. The Set-Cookie expires field takes JavaScript Date
object [1].

Fix this issue by converting redisplay UTC time to local time. Also,
convert redisplay from Date to Timestamp in the REST API, since
universally supported format for JavaScript's Date object is
ISO 8601 [2] and it is easy to convert Timestamp to this format.
Also, this is the standard approach used in Gerrit core for displaying
date and time on web UI.

[1] https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Set-Cookie
[2] https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date#date_time_string_format

Change-Id: Ic24b758a6f0d807c0fa6a2ed3f9bda125fa31c60