Allow reading message content and config from a repo
In a multi-primary Gerrit setup, updating the content and expiration
of the message on each primary can be challenging. Update the plugin
to allow reading the message and its configuration from a repository.
The advantage of storing these in a repository is that any updates
will be automatically reflected across all primaries.

Change-Id: Idb6f660aa79dc9d836244c10054a0cd02e078464