Show banner after dismissal if message is updated
The banner is currently not displayed to the user if they dismiss it
until the next day even if the message was updated. When the message
content is updated, the intent is usually to show it to everyone, even
those who dismissed the previous message. This is achieved by updating
the REST API to return the 'id' based on the message content, as the
cookie to manage dismissals is named using this 'id'.

Change-Id: I2ccaaa55ac549cb37eed851f8ed40f4d5f345f72