Implement Git based message store
With this feature we can use Gerrit UI for creating and reviewing
message of the day.

To use Git based config, set this in gerrit.config:
[plugin "messageoftheday"]
  configGit = motd

Then create messageoftheday.config and the corresponding html message
files in the master branch of the motd repository.

Setting message of the day via REST API is also supported for the Git
based message store but I am not sure that this was worth the effort
considering that we can use both direct Git push and Gerrit UI for
creating and maintaining the message of the day.

Change-Id: I7eda83cce8745dad6015a6856d521afb8221c016