Fix date/time parsing for the set-message REST API
When the locale is not specified the parsing could fail because of
subtle differences in how AM/PM is given in different locale(s). For
example, in Germany the UI sent a timestamp with uppercase 'PM' but the
DateTimeFormatter created with German locale expected lowercase 'pm'.

Set the locale to ENGLISH to make parsing behaviour predictable.

Change-Id: I82de7cde385251472853ab8998a5237268234187